backend/requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
